+++ Sources/Web3Auth/KeychainManager.swift
@@ -5,7 +5,7 @@
 //  Created by Dhruv Jaiswal on 18/07/22.
 //
 
-import Foundation
+
 import KeychainSwift
 
 enum KeychainConstantEnum {
