+++ Sources/Web3Auth/SECP256k1.swift
@@ -1,6 +1,7 @@
 import CryptoSwift
 import Foundation
 import secp256k1
+
 public struct SECP256K1 {
     public struct UnmarshaledSignature {
         public var v: UInt8 = 0
@@ -15,27 +16,10 @@ public struct SECP256K1 {
     }
 }
 
-extension secp256k1.Signing.ECDSASigner {
-    public func signatureKeccaf256Hash(for data: Data) throws -> secp256k1.Signing.ECDSASignature {
-        try signatureKeccaf256(for: data.sha3(.keccak256))
-    }
-}
 
 extension SECP256K1 {
     static let context = secp256k1_context_create(UInt32(SECP256K1_CONTEXT_SIGN | SECP256K1_CONTEXT_VERIFY))
 
-    func sign(privkey: String, messageData: String) -> String {
-        let privateBytes = privkey.hexa
-        let privateKey = try! secp256k1.Signing.PrivateKey(rawRepresentation: privateBytes)
-        //  Public key
-        print(String(bytes: privateKey.publicKey.rawRepresentation))
-
-        // ECDSA
-        let messageData = messageData.data(using: .utf8)!
-        let signature = try! privateKey.ecdsa.signatureKeccaf256Hash(for: messageData)
-
-        return (try! signature.derRepresentation.toHexString())
-    }
 
     public static func ecdh(pubKey: secp256k1_pubkey, privateKey: Data) -> secp256k1_pubkey? {
         var localPubkey = pubKey // Pointer takes a variable
